Updated icons, added pointer

+++ PeTedit/NetItems/PetriNetScene.cpp
@@ -419,7 +419,7 @@ void PetriNetScene::produce(PNMLBuilder* builder){
 	produce((PetriEngine::AbstractPetriNetBuilder*)builder);
 
 	//Output queries
-	for(size_t i = 0; i < _queries->rowCount(); i++)
+	for(int i = 0; i < _queries->rowCount(); i++)
 		builder->addQuery(_queries->query(i));
 }
